home_renovation_assistant/requirements.txt
streamlit==1.45.0
pydantic-ai-slim[openai,logfire]==0.1.8
pydantic==2.11.4
python-dotenv==1.0.0
openai==1.75.0
pandas==2.2.3
pillow==11.2.1
langfuse==2.60.3
nest-asyncio==1.6.0
logfire==3.14.1